Move custom type to manimlib.typing
