+++ manimlib/mobject/frame.py
@@ -6,7 +6,7 @@ from manimlib.mobject.geometry import Rectangle
 
 from typing import TYPE_CHECKING
 if TYPE_CHECKING:
-    from manimlib.constants import ManimColor
+    from manimlib.typing import ManimColor
 
 
 class ScreenRectangle(Rectangle):
