+++ manimlib/mobject/svg/drawings.py
@@ -60,7 +60,7 @@ from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from typing import Tuple, Sequence, Callable
-    from manimlib.constants import ManimColor, np_vector
+    from manimlib.typing import ManimColor, np_vector
 
 
 class Checkmark(TexTextFromPresetString):
