+++ manimlib/mobject/svg/mtex_mobject.py
@@ -13,7 +13,7 @@ if TYPE_CHECKING:
     from typing import Iterable, Union
 
     from manimlib.mobject.types.vectorized_mobject import VGroup
-    from manimlib.constants import ManimColor
+    from manimlib.typing import ManimColor
 
     Span = tuple[int, int]
     Selector = Union[
