+++ manimlib/mobject/svg/svg_mobject.py
@@ -22,7 +22,7 @@ from manimlib.utils.simple_functions import hash_string
 
 from typing import TYPE_CHECKING
 if TYPE_CHECKING:
-    from manimlib.constants import ManimColor
+    from manimlib.typing import ManimColor
