+++ manimlib/mobject/types/point_cloud_mobject.py
@@ -15,7 +15,7 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from typing import Callable, Sequence
     import numpy.typing as npt
-    from manimlib.constants import ManimColor, np_vector
+    from manimlib.typing import ManimColor, np_vector
 
 
 class PMobject(Mobject):
